third_party/WebKit/LayoutTests/http/tests/performance-timing/longtask-v2/longtask-v8compile.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>LongTask Timing: validate the long task attributes for script-compile</title>
<body>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<h1>Long Task Attributes for Script-compile</h1>
<div id="log"></div>
<script>
  async_test(function (t) {
    const scriptURL = 'resources/makelongtask.js';
    const lineColumn = '(0, 0)';
    const observer = new PerformanceObserver(
      t.step_func(function(entryList) {
        const longtaskEntries = entryList.getEntries().filter((e) => {
          return e.attribution.find((a) => {
            return a.scriptURL.includes(scriptURL);
          });
        });
        if (longtaskEntries.length === 0) {
          return;
        }
        assert_equals(longtaskEntries.length, 1);
        const longtask = longtaskEntries[0];
        assert_equals(longtask.entryType, 'longtask');
        assert_equals(longtask.name, 'self');
        assert_greater_than(longtask.duration, 50);
        assert_greater_than_equal(longtask.startTime, 0);
        const currentTime = performance.now();
        assert_less_than_equal(longtask.startTime, currentTime);

        const compileAttribution = longtask.attribution.filter((it) => {
          return it.name === 'script-compile' &&
              it.scriptURL.includes(scriptURL);
        });
        assert_equals(compileAttribution.length, 1);
        const it = compileAttribution[0];
        assert_greater_than(it.startTime, 0);
        assert_equals(it.entryType, 'taskattribution');
        assert_equals(it.containerType, 'iframe');
        assert_equals(it.containerId, '');
        assert_equals(it.containerName, '');
        assert_equals(it.containerSrc, '');
        assert_true(it.scriptURL.includes(scriptURL + lineColumn));
        observer.disconnect();
        t.done();
      })
    );
    observer.observe({entryTypes: ['longtask']});

    internals.BypassLongCompileThresholdOnce();

    var iframe=document.createElement('Script');
    iframe.src = "resources/makelongtask.js";
    document.body.appendChild(iframe);

}, 'Performance longtask entries with script-compile attribute are observable.');
</script>

</body>
