third_party/WebKit/LayoutTests/virtual/longtask-v2/http/tests/performance-timing/longtask-v2/longtask-v8compile-expected.txt

This is a testharness.js-based test.
FAIL Performance longtask entries with script-compile attribute are observable. assert_equals: expected "self" but got "unknown"
Harness: the test ran to completion.
